requirements.txt
cfgv==3.3.1
cryptography==42.0.4
PyJWT==2.8.0
requests==2.31.0
rfc3986==2.0.0
websocket-client==1.8.0